changed getSessionTimeout to return a long.

+++ cpp/src/Glacier2/RouterI.cpp
@@ -171,7 +171,7 @@ Glacier2::RouterI::destroySession(const Current&)
     assert(false); // Must not be called in this router implementation.
 }
 
-Ice::Int
+Ice::Long
 Glacier2::RouterI::getSessionTimeout(const Current&) const
 {
     assert(false); // Must not be called in this router implementation.
